Added pulltweets. help and cardslist
Pulling tweets from twitter account is possible now. Added help which is just a template for now until I update the documentation. The cards command now pulls information from the cardsList.json will be write in this file soon too for people to be able to edit their own profile soon since the command is tied to the id of user